Add files via upload
moved fcfs to external file
